+++ app/routes.js
@@ -8,7 +8,7 @@ import { withCatch, getHealthCheckDependencies, extractAccessToken } from './hel
 import { getDashboardContext } from './pages/dashboard/controller';
 import { getDescriptionContext, getDescriptionErrorContext, postOrPutDescription } from './pages/sections/description/controller';
 import includesContext from './includes/manifest.json';
-import { getTaskListPageContext } from './pages/order-task-list/controller';
+import { getTaskListPageContext } from './pages/task-list/controller';
 
 const addContext = ({ context, user, csrfToken }) => ({
   ...context,
@@ -46,7 +46,7 @@ export const routes = (authProvider) => {
     const accessToken = extractAccessToken({ req, tokenType: 'access' });
     const { orderId } = req.params;
     const context = await getTaskListPageContext({ accessToken, orderId });
-    res.render('pages/order-task-list/template.njk', addContext({ context, user: req.user }));
+    res.render('pages/task-list/template.njk', addContext({ context, user: req.user }));
   }));
 
   router.get('/organisation/:orderId/description', authProvider.authorise({ claim: 'ordering' }), withCatch(authProvider, async (req, res) => {
